requirements.txt
gunicorn
psycopg2-binary
python-decouple
dj-database-url
whitenoise
Pillow  
djangorestframework
